+++ src/mlpack/bindings/markdown/print_docs.cpp
@@ -9,11 +9,13 @@
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#include "print_docs.hpp"
-
 #include <mlpack/core/util/io.hpp>
 #include <mlpack/core/util/binding_details.hpp>
+
+#include <boost/algorithm/string/replace.hpp>
+
 #include "binding_info.hpp"
+#include "print_docs.hpp"
 #include "print_doc_functions.hpp"
 
 // Make sure that this is defined.