+++ src/mlpack/methods/ann/layer/batch_norm_impl.hpp
@@ -234,7 +234,7 @@ void BatchNorm<InputDataType, OutputDataType>::serialize(
   if (cereal::is_loading<Archive>())
   {
     weights.set_size(size + size, 1);
-    loading = false;
+    loading = true;
   }
 
   ar(CEREAL_NVP(eps));